+++ src/bin/rage-keygen/main.rs
@@ -35,7 +35,7 @@ fn main() {
             "# created: {}",
             chrono::Local::now().to_rfc3339_opts(chrono::SecondsFormat::Secs, true)
         )?;
-        writeln!(output, "# public key: {}", sk.to_public().to_str())?;
+        writeln!(output, "# public key: {}", sk.to_public())?;
         writeln!(output, "{}", sk.to_string().expose_secret())
     })() {
         error!("Failed to write to output: {}", e);